+++ Source/ACE.DatLoader/Entity/AnimationHooks/IHook.cs
@@ -1,0 +1,6 @@
+﻿namespace ACE.DatLoader.Entity
+{
+    public interface IHook
+    {
+    }
+}