+++ Source/ACE.DatLoader/Entity/AnimationHooks/SoundTableHook.cs
@@ -1,0 +1,14 @@
+﻿namespace ACE.DatLoader.Entity.AnimationHooks
+{
+    public class SoundTableHook : IHook
+    {
+        public uint SoundType { get; set; }
+
+        public static SoundTableHook ReadHookType(DatReader datReader)
+        {
+            SoundTableHook s = new SoundTableHook();
+            s.SoundType = datReader.ReadUInt32();
+            return s;
+        }
+    }
+}